+++ refinedstorage2-network-api/src/main/java/com/refinedmods/refinedstorage2/api/network/node/exporter/package-info.java
@@ -1,6 +1,6 @@
 @ParametersAreNonnullByDefault
 @FieldsAndMethodsAreNonnullByDefault
-package com.refinedmods.refinedstorage2.api.network.node.exporter.strategy;
+package com.refinedmods.refinedstorage2.api.network.node.exporter;
 
 import com.refinedmods.refinedstorage2.api.core.FieldsAndMethodsAreNonnullByDefault;