+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/ModInitializer.java
@@ -221,7 +221,7 @@ public class ModInitializer extends AbstractModInitializer {
 
     private void registerGridExtractionStrategyFactories() {
         PlatformApi.INSTANCE.addGridExtractionStrategyFactory(
-            (containerMenu, player, gridServiceFactory, containerExtractionSource) ->
+            (containerMenu, player, gridServiceFactory, itemStorage) ->
                 new ItemGridExtractionStrategy(containerMenu, player, gridServiceFactory)
         );
         PlatformApi.INSTANCE.addGridExtractionStrategyFactory(FluidGridExtractionStrategy::new);